feat: Implement category and product services with CRUD operations and update related DTOs and controllers

+++ Api.ViewModel/DTOs/LigneCommandeDtos.cs
@@ -2,13 +2,13 @@ namespace Api.ViewModel.DTOs;
 
 public record LigneCommandeCreateDto(
     int Quantite,
-    decimal PrixUnitaire,
+    float PrixUnitaire,
     int CommandeId,
     int ProduitId
 );
 
 public record LigneCommandeUpdateDto(
     int Quantite,
-    decimal PrixUnitaire,
+    float PrixUnitaire,
     int ProduitId
 );
